requirements.txt
langchain==0.0.331
tiktoken==0.5.2
openai==0.28.1
pinecone-client==2.2.4
Flask==3.0.0
Flask-Cors==4.0.0
gradio